pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "econdatapy"
version = "0.1.0"
authors = [
  { name="Byron Botha", email="byron@codera.co.za" },
]
description = "A package for downloading data from www.econdata.co.za"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pandas",
  "pyjwt",
  "requests"
]
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/coderaanalytics/econdatapy"
Issues = "https://github.com/coderaanalytics/econdatapy/issues"
